+++ R/imap.R
@@ -49,6 +49,13 @@ imap_dbl <- function(.x, .f, ...) {
 
 #' @rdname imap
 #' @export
+imap_raw <- function(.x, .f, ...) {
+  .f <- as_mapper(.f, ...)
+  map2_raw(.x, vec_index(.x), .f, ...)
+}
+
+#' @rdname imap
+#' @export
 imap_dfr <- function(.x, .f, ..., .id = NULL) {
   .f <- as_mapper(.f, ...)
   map2_dfr(.x, vec_index(.x), .f, ..., .id = .id)