+++ tests/testthat/test-as-mapper.R
@@ -71,7 +71,10 @@ test_that("lists are wrapped", {
   expect_identical(mapper_list, base_list)
 })
 
-test_that("raw and complex types aren't supported for indexing", {
-  expect_error(as_mapper(1)(raw(2)))
+test_that("complex types aren't supported for indexing", {
   expect_error(as_mapper(1)(complex(2)))
+})
+
+test_that("raw vectors are supported for indexing", {
+  expect_equal( as_mapper(1)(raw(2)), raw(1) )
 })