+++ NEWS.md
@@ -22,6 +22,9 @@
 
 * `list_modify()`, `update_list()` and `list_merge()` now handle duplicate
   duplicate argument names correctly (#441, @mgirlich).
+  
+* `map_raw`, `imap_raw`, `flatten_raw`, `invoke_map_raw`, `map2_raw` and `pmap_raw`
+  added to support raw vectors. (#455, @romainfrancois)
 
 # purrr 0.2.4